requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_tflow_select=2.3.0=mkl
absl-py=0.15.0=pyhd8ed1ab_0
anyio=3.5.0=py37h89c1867_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py37h5e8e339_1
astor=0.8.1=pyh9f0ad1d_0
attrs=21.4.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.10.0=pyha770c72_0
blas=2.17=openblas
bleach=4.1.0=pyhd8ed1ab_0
brotlipy=0.7.0=py37h5e8e339_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cairo=1.16.0=hf32fb01_1
certifi=2021.10.8=py37h89c1867_2
cffi=1.15.0=py37h036bc23_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
cryptography=36.0.0=py37h9ce1e76_0
cycler=0.11.0=pypi_0
dbus=1.13.6=he372182_0
debugpy=1.5.1=py37hcd2ae1e_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
eigen=3.3.9=h4bd325d_1
entrypoints=0.4=pyhd8ed1ab_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.0=hcdf2ecd_0
flit-core=3.7.1=pyhd8ed1ab_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.31.2=pypi_0
freeglut=3.2.1=h9c3ff4c_2
freetype=2.10.4=h0708190_1
gast=0.2.2=py_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
glib=2.70.2=h780b84a_4
glib-tools=2.70.2=h780b84a_4
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.14=h23475e2_0
grpcio=1.45.0=py37h0327239_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.8.0=py37h3010b51_1003
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
icu=58.2=hf484d3e_1000
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=py37h89c1867_1
importlib_resources=5.6.0=pyhd8ed1ab_0
ipykernel=6.11.0=py37h25bab4e_0
ipython=7.32.0=py37h89c1867_0
ipython_genutils=0.2.0=py_1
jasper=2.0.33=ha77e612_0
jedi=0.18.1=py37h89c1867_0
jinja2=3.1.1=pyhd8ed1ab_0
jpeg=9d=h7f8727e_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter_client=7.2.1=pyhd8ed1ab_0
jupyter_core=4.9.2=py37h89c1867_0
jupyter_server=1.16.0=pyhd8ed1ab_1
jupyterlab=3.3.2=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.12.0=pyhd8ed1ab_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.4.2=pypi_0
lame=3.100=h7f98852_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.8.0=17_openblas
libcblas=3.8.0=17_openblas
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_14
libgfortran=3.0.0=1
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libglib=2.70.2=h174f98d_4
libglu=9.0.0=he1b5a44_1001
libgomp=11.2.0=h1d223b6_14
libiconv=1.16=h516909a_0
liblapack=3.8.0=17_openblas
liblapacke=3.8.0=17_openblas
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.10=h5a2b251_0
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.20.0=h6239696_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=he4da1e4_14
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
libzlib=1.2.11=h166bdaf_1014
llvm-openmp=13.0.1=he0ac6c6_1
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.1.1=py37h540881e_1
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py37h5e8e339_1005
nbclassic=0.3.7=pyhd8ed1ab_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.5=pyhd8ed1ab_1
nbconvert-core=6.4.5=pyhd8ed1ab_1
nbconvert-pandoc=6.4.5=pyhd8ed1ab_1
nbformat=5.2.0=pyhd8ed1ab_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
notebook=6.4.10=pyha770c72_0
notebook-shim=0.1.0=pyhd8ed1ab_0
numpy=1.21.5=pypi_0
olefile=0.46=py37_0
openblas=0.3.3=h9ac9557_1001
opencv=3.4.2=py37h6fd60c2_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=3.0.2=h166bdaf_1
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd8ed1ab_0
pandoc=2.17.1.1=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.0.0=py37h9a89aac_0
pip=21.2.2=py37h06a4308_0
pixman=0.40.0=h36c2ea0_0
prometheus_client=0.13.1=pyhd8ed1ab_0
prompt-toolkit=3.0.27=pyha770c72_0
protobuf=3.20.0=py37hd23a5d3_0
psutil=5.9.0=py37h5e8e339_0
ptyprocess=0.7.0=pyhd3deb0d_0
py-opencv=3.4.2=py37hb342d67_1
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyrsistent=0.18.1=py37h5e8e339_0
pysocks=1.7.1=py37h89c1867_4
python=3.7.12=hf930737_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2022.1=pyhd8ed1ab_0
pyzmq=22.3.0=py37h336d617_1
qt=5.6.3=h8bf5577_3
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd8ed1ab_0
scipy=1.5.3=py37h8911b10_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=61.3.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py37h89c1867_2
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.2=hc218d9a_0
tensorboard=1.15.0=py37_0
tensorflow=1.15.0=mkl_py37h28c19af_0
tensorflow-base=1.15.0=mkl_py37he1670d9_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py_2
terminado=0.13.3=py37h89c1867_0
testpath=0.6.0=pyhd8ed1ab_0
tk=8.6.11=h1ccaba5_0
tornado=6.1=py37h5e8e339_2
traitlets=5.1.1=pyhd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
urllib3=1.26.9=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.3.2=pyhd8ed1ab_0
werkzeug=0.16.1=py_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.0=py37h540881e_0
x264=1!157.20191217=h7b6447c_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h166bdaf_1014
zstd=1.4.9=ha95c52a_0
